+++ .idea/misc.xml
@@ -47,6 +47,7 @@
         <entry key="app/src/main/res/layout/fragment_eat.xml" value="0.15577507598784193" />
         <entry key="app/src/main/res/layout/fragment_event_details.xml" value="0.375" />
         <entry key="app/src/main/res/layout/fragment_events.xml" value="0.25" />
+        <entry key="app/src/main/res/layout/fragment_filters.xml" value="0.5631067961165048" />
         <entry key="app/src/main/res/layout/fragment_meet.xml" value="0.4461669921875" />
         <entry key="app/src/main/res/layout/fragment_meetOLD.xml" value="0.24326635314236852" />
         <entry key="app/src/main/res/layout/fragment_more.xml" value="0.25" />
