+++ app/src/main/java/it/bz/noi/community/ui/today/news/NewsPagingSource.kt
@@ -1,4 +1,4 @@
-package it.bz.noi.community.ui.today
+package it.bz.noi.community.ui.today.news
 
 import android.util.Log
 import androidx.paging.PagingSource
